requirements.txt
Package           Version
----------------- ---------
certifi           2020.12.5
chardet           4.0.0
colorama          0.4.4
configparser      5.0.1
crayons           0.4.0
et-xmlfile        1.0.1
idna              2.10
jdcal             1.4.1
numpy             1.19.5
openpyxl          3.0.6
pandas            1.2.0
Pillow            8.1.0
pip               20.3.3
python-dateutil   2.8.1
pytz              2020.5
requests          2.25.1
selenium          3.141.0
setuptools        51.0.0
six               1.15.0
urllib3           1.26.2
webdriver-manager 3.2.2
wheel             0.36.2
xlrd              2.0.1
